+++ sdk/clojure/src/test/core-sdk/starfederation/datastar/clojure/adapter/common_test.clj
@@ -52,7 +52,7 @@
 ;; -----------------------------------------------------------------------------
 (defn ->machinery [write-profile]
   (let [^ByteArrayOutputStream baos (ByteArrayOutputStream.)
-        {write! ac/write-to-buffered-writer!
+        {write! ac/write!
          wrap   ac/wrap-output-stream} write-profile
         writer (wrap baos)]
     {:write! write!